Better diffl()
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/turbine/trunk@9713 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/tcl/adlb/tcl-adlb.h
@@ -26,7 +26,7 @@ extern int ADLB_curr_priority;
 extern int adlb_comm_rank;
 
 /* Return a pointer to a shared buffer */
-char *tcl_adlb_xfer_buffer(int *buf_size);
+char *tcl_adlb_xfer_buffer(uint64_t *buf_size);
 
 int type_from_obj(Tcl_Interp *interp, Tcl_Obj *const objv[],
                          Tcl_Obj* obj, adlb_data_type *type);
